pyproject.toml
[tool.poetry]
name = "python-etl-duckdb"
version = "0.1.0"
description = ""
authors = ["Thiago"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.32.1"
gdown = "^5.1.0"
duckdb = "^0.10.0"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.28"
python-dotenv = "^1.0.1"
pandas = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
